Fix latent reconfigure bugs and clean up code quality
- pwr_dims_equal: detect range_start / bin-spacing changes so the
  lightweight path never runs on a stale decimation grid
- recalibrate the velocity axis on carrier-only reconfiguration
- ui_theme: fix UI_C_SERIES_3 channel truncation (0x8BC -> 8B,C3,4A)
- deduplicate STC ramp and UI_PI definitions, remove dead code,
  fix stale comments, document the wheel==2 double-click convention

+++ PWRadarCore/src/pwr_cfar.c
@@ -590,9 +590,9 @@ void pwr_cluster_run(struct PWR_Engine* e)
                                        cen_d * e->axis_vel_step_mps;
             det->azimuth_deg  = pwr_wrap360(e->beam_azimuth_deg);
             det->time_s       = e->scenario_time_s;
-            det->amplitude_db = (float)peak_db;
-            det->snr_db       = (float)snr_db;
-            det->threshold_db = (float)pwr_pow_to_db(
+            det->amplitude_db = peak_db;
+            det->snr_db       = snr_db;
+            det->threshold_db = (double)pwr_pow_to_db(
                 e->cfar.threshold[(size_t)acc.peak_j * nr + acc.peak_r]);
             det->ambiguous = (det->range_m > e->dm.unambiguous_range_m ||
                               fabs(det->radial_velocity_mps) >
